+++ src/asyncclick/termui.py
@@ -93,6 +93,7 @@ async def prompt(
     show_default: bool = True,
     err: bool = False,
     show_choices: bool = True,
+    blocking: bool = True,
 ) -> t.Any:
     """Prompts a user for input.  This is a convenience function that can
     be used to prompt a user for input later.
@@ -120,10 +121,17 @@ async def prompt(
                          For example if type is a Choice of either day or week,
                          show_choices is true and text is "Group by" then the
                          prompt will be "Group by (day, week): ".
+    :param blocking: if `False`, uses a thread to interact with the terminal.
+                     This keeps the event loop running but may cause interesting
+                     effects when interrupted with Control-C. The default
+                     is `True`, but don't depend on it.
 
     Warning: The user interaction is run inside a separate thread, because otherwise
     the call would block the async loop. As a result, behavior when interrupted may be
     sub-optimal.
+
+    .. versionadded:: 8.3
+        (AsyncClick) ``prompt`` is now async; ``blocking`` parameter added.
 
     .. versionadded:: 8.0
         ``confirmation_prompt`` can be a custom string.
@@ -156,6 +164,13 @@ async def prompt(
                 echo(None, err=err)
             raise Abort() from None
 
+    if blocking:
+        async def run_prompt_func(text: str) -> str:
+            return prompt_func(text)
+    else:
+        def run_prompt_func(text: str) -> Awaitable[str]:
+            return anyio.to_thread.run_sync(prompt_func, text)
+
     if value_proc is None:
         value_proc = convert_type(type, default)
 
@@ -171,7 +186,7 @@ async def prompt(
 
     while True:
         while True:
-            value = await anyio.to_thread.run_sync(prompt_func, prompt)
+            value = await run_prompt_func(prompt)
             if value:
                 break
             elif default is not None:
@@ -190,7 +205,7 @@ async def prompt(
         if not confirmation_prompt:
             return result
         while True:
-            value2 = await anyio.to_thread.run_sync(prompt_func, confirmation_prompt)
+            value2 = await run_prompt_func(confirmation_prompt)
             is_empty = not value and not value2
             if value2 or is_empty:
                 break